requirements.txt
streamlit
pandas
llama-cpp-python==0.2.24
numpy
scikit-learn
fpdf
